fix(security): float→Decimal for consensus/agent/onboarding/genesis ECU interfaces
consensus/governance.py: ecu_base_costs Dict[str, Decimal]; get_task_fee_ecu() → Decimal. consensus/engine.py: node_stakes Dict[str, Decimal]. agent.py: stake/reward method signatures → Decimal. onboarding.py: credit/repayment → Decimal. exceptions.py: stake/required → Decimal. genesis/work_task.py: ecu_estimate → Optional[Decimal]. Directive 3 compliance per ILC Coding Security Standards (CLAUDE.md). Identified Phase 1117 A11 audit.

float_kill_01_interface_decimal_cleanup

+++ ilc_core/exceptions.py
@@ -2,6 +2,8 @@
 
 from __future__ import annotations
 
+from decimal import Decimal
+
 
 class IlcError(Exception):
     """Base exception for all ILC domain errors."""
@@ -13,8 +15,8 @@ class InsufficientStakeError(IlcError, ValueError):
     def __init__(
         self,
         agent_id: str,
-        stake: float,
-        required: float,
+        stake: Decimal,
+        required: Decimal,
         message: str = "",
     ) -> None:
         self.agent_id = agent_id
